1268094: Avoid traceback on unreg with >1 sub
When unregistering with the gui, if there are more
than one subscriptions attached, there is a traceback
on the console.

The on_selection callback is getting called with a
selection that points to a serial/cert that isn't
valid by the time the callback runs. So the the
cert lookup returns None, and the rest of the code
didn't expect that. So just check for that and return.